backend/requirements.txt
bcrypt
celery
flasgger
flask
flask-cors
flask-migrate
flask-sqlalchemy
pyjwt
python-dotenv
redis